Implemented basic auth check upon refresh and redefined routes
