+++ src/App.js
@@ -9,22 +9,25 @@ import FlashMessagesList from './components/flash/FlashMessagesList';
 import requireAuth from './utils/requireAuth';
 import './styles/App.css';
 
+import Profile from './components/Profile';
+
 class App extends React.Component {
-    render() {
-        return (
-          <div>
-            
-            <NavBar />
-            <FlashMessagesList />
-            
-            <Route exact path="/" component={HomePage} />
-            <Route path="/login" component={LoginPage} />
-            <Route path="/verify_account" component={PassportPage} />
-            <Route path="/dashboard" component={requireAuth(Dashboard)} />
-            
-          </div>
-        );
-    }
+  render() {
+    return (
+      <div>
+
+        <NavBar />
+        <FlashMessagesList />
+
+        <Route exact path="/" component={HomePage} />
+        <Route path="/login" component={LoginPage} />
+        <Route path="/verify_account" component={PassportPage} />
+
+        <Route path="/dashboard" component={Dashboard} />
+
+      </div>
+    );
+  }
 }
 
 export default App;
