+++ src/components/Account.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class Account extends React.Component {
+  render() {
+    return (
+      <div>
+        <h1>This is the Account component.</h1>
+      </div>
+    );
+  }
+};
