+++ src/components/Community.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class Community extends React.Component {
+  render() {
+    return (
+      <div>
+        <h1>This is the Community component.</h1>
+      </div>
+    );
+  }
+};
