+++ src/reducers/user.js
@@ -1,11 +1,25 @@
 import { SAVE_USER } from '../actions/types';
 
-export default (state = {}, action = {}) => {
+const defaultUser = {
+  username: '',
+  email: '',
+  avatarUrl: '',
+  githubId: '',
+  githubData: {},
+  fccCerts: {},
+  skills: [],
+  projects: [],
+  interests: [],
+  memberSince: '',
+  verifiedUser: false
+}
+
+export default (state = defaultUser, action) => {
   switch (action.type) {
-    
-    case SAVE_USER: 
+
+    case SAVE_USER:
       return action.user;
-      
+
     default: return state;
   }
 }
